refactor: health controller
