+++ packages/server/src/routes/health/__tests__/health.test.ts
@@ -2,7 +2,7 @@ import app from '../../../app'
 import request from 'supertest'
 
 describe('GET /health', () => {
-  it('response health controller', async () => {
+  it('response health ', async () => {
     const response = await request(app).get('/health')
 
     expect(response.statusCode).toBe(200)
